LayoutTests/platform/chromium/fast/forms/calendar-picker/week-picker-mouse-operations-expected.txt

Tests if week picker mouse operations work as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that page popup doesn't exist at first.
PASS document.getElementById("mock-page-popup") is null
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W01"
PASS highlightedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
Check that hovering over an entry highlights it.
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "1999-W51"
PASS highlightedDayCells() is "1999-12-26"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "1999-W52"
PASS highlightedDayCells() is "1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31,2000-01-01,2000-01-02"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W01"
PASS highlightedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS currentMonth() is "2000-01"
PASS selectedValue() is "2000-W01"
PASS selectedDayCells() is "2000-01-03,2000-01-04,2000-01-05,2000-01-06,2000-01-07,2000-01-08,2000-01-09"
PASS highlightedValue() is "2000-W05"
PASS highlightedDayCells() is "2000-01-31,2000-02-01,2000-02-02,2000-02-03,2000-02-04,2000-02-05"
Check that mouse click sets the value.
PASS currentMonth() is "2000-01"
PASS selectedValue() is "1999-W52"
PASS selectedDayCells() is "1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31,2000-01-01,2000-01-02"
PASS highlightedValue() is "1999-W52"
PASS highlightedDayCells() is "1999-12-27,1999-12-28,1999-12-29,1999-12-30,1999-12-31,2000-01-01,2000-01-02"
PASS document.getElementById("week").value is "1999-W52"
PASS successfullyParsed is true

TEST COMPLETE
 
